+++ db/updates/update_10.4-10.5.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `documents` ADD `isSDS` INT NOT NULL DEFAULT '0' AFTER `isBatch`;