+++ releasenotes.md
@@ -1,4 +1,7 @@
 Whats New in v10.5
 --------------------------
+- Introduse SDS creation (PREVIEW)
+- SDS in ingredient has been renamed to a Document
+- Added synonyms in document generation for ingredients
 - Auto create ingredient when importing a formula from Marketplace if not exists locally 
 - For full details please refer to the CHANGELOG